Conflicts:
	admin/broadleaf-admin-module/pom.xml
	admin/broadleaf-contentmanagement-module/pom.xml
	admin/broadleaf-open-admin-platform/pom.xml
	admin/pom.xml
	common/pom.xml
	core/broadleaf-framework-web/pom.xml
	core/broadleaf-framework/pom.xml
	core/broadleaf-profile-web/pom.xml
	core/broadleaf-profile/pom.xml
	core/pom.xml
	integration/pom.xml
	pom.xml

+++ core/broadleaf-framework/src/main/java/org/broadleafcommerce/core/order/service/OrderServiceImpl.java
@@ -54,6 +54,7 @@ import org.springframework.jmx.export.annotation.ManagedAttribute;
 import org.springframework.jmx.export.annotation.ManagedResource;
 import org.springframework.orm.jpa.JpaTransactionManager;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.PlatformTransactionManager;
 import org.springframework.transaction.TransactionDefinition;
 import org.springframework.transaction.TransactionStatus;
 import org.springframework.transaction.annotation.Transactional;
@@ -118,7 +119,7 @@ public class OrderServiceImpl implements OrderService {
     protected SequenceProcessor removeItemWorkflow;
 
     @Resource(name = "blTransactionManager")
-    JpaTransactionManager transactionManager;
+    protected PlatformTransactionManager transactionManager;
 
     @Value("${pricing.retry.count.for.lock.failure}")
     protected int pricingRetryCountForLockFailure = 3;
@@ -590,6 +591,28 @@ public class OrderServiceImpl implements OrderService {
             }
             order.getPaymentInfos().remove(paymentInfo);
             paymentInfo = paymentInfoDao.readPaymentInfoById(paymentInfo.getId());
+            paymentInfoDao.delete(paymentInfo);
+        }
+    }
+
+    @Override
+    @Transactional("blTransactionManager")
+    public void removePaymentFromOrder(Order order, PaymentInfo paymentInfo){
+        PaymentInfo paymentInfoToRemove = null;
+        for(PaymentInfo info : order.getPaymentInfos()){
+            if(info.equals(paymentInfo)){
+                paymentInfoToRemove = info;
+            }
+        }
+        if(paymentInfoToRemove != null){
+            try {
+                securePaymentInfoService.findAndRemoveSecurePaymentInfo(paymentInfoToRemove.getReferenceNumber(), paymentInfo.getType());
+            } catch (WorkflowException e) {
+                // do nothing--this is an acceptable condition
+                LOG.debug("No secure payment is associated with the PaymentInfo", e);
+            }
+            order.getPaymentInfos().remove(paymentInfoToRemove);
+            paymentInfo = paymentInfoDao.readPaymentInfoById(paymentInfoToRemove.getId());
             paymentInfoDao.delete(paymentInfo);
         }
     }
